requirements.txt
langgraph
langchain
langchain-openai
pydantic
azure-devops
azure-functions
python-dotenv